# src/main/java/com/SmartMarket/dto/AuthRequest.java
package com.SmartMarket.dto;

import jakarta.validation.constraints.NotEmpty;
import jakarta.validation.constraints.NotNull;

public class AuthRequest {
	@NotNull
    @NotEmpty
    private String username;
	@NotNull
    @NotEmpty
    private String password;
	
    
	public String getUsername() {
		return username;
	}
	public void setUsername(String username) {
		this.username = username;
	}
	public String getPassword() {
		return password;
	}
	public void setPassword(String password) {
		this.password = password;
	}

    // Getter ve Setter
}