# src/main/java/com/SmartMarket/dto/AuthResponse.java
package com.SmartMarket.dto;

import jakarta.validation.constraints.NotEmpty;
import jakarta.validation.constraints.NotNull;

public class AuthResponse {
	@NotNull
    @NotEmpty
    private String token;

    public AuthResponse(String token) {
        this.token = token;
    }

    public String getToken() {
        return token;
    }
}