# src/main/resources/static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SmartMarket API",
    "description": "Do'kon mahsulotlari va savdo API'si - Store ID token'dan otomatik alınır",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8081"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "Tizimga kirish",
        "tags": ["Authentication"],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Muvaffaqiyatli kirish",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/products": {
      "post": {
        "summary": "Mahsulot qo'shish",
        "tags": ["Products"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductsObject"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mahsulot muvaffaqiyatli qo'shildi"
          }
        }
      },
      "put": {
        "summary": "Mahsulot yangilash",
        "tags": ["Products"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductsObject"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mahsulot yangilandi"
          }
        }
      },
      "delete": {
        "summary": "Mahsulot o'chirish",
        "tags": ["Products"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductsObject"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mahsulot o'chirildi"
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "summary": "Mahsulot olish",
        "description": "Store ID token'dan otomatik alınır",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Mahsulot ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Mahsulot ma'lumotlari",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductsObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/stores": {
      "get": {
        "summary": "Do'kon ma'lumotlarini olish",
        "description": "Store ID token'dan otomatik alınır",
        "tags": ["Stores"],
        "responses": {
          "200": {
            "description": "Do'kon ma'lumotlari",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Do'kon yangilash",
        "tags": ["Stores"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Do'kon muvaffaqiyatli yangilandi"
          }
        }
      }
    },
    "/api/stores/{id}/products": {
      "get": {
        "summary": "Do'kon mahsulotlarini olish",
        "description": "Store ID token'dan otomatik alınır",
        "tags": ["Products"],
        "responses": {
          "200": {
            "description": "Mahsulotlar ro'yxati",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductsObject"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/stores/password": {
      "put": {
        "summary": "Do'kon parolini yangilash",
        "description": "Store ID token'dan otomatik alınır",
        "tags": ["Stores"],
        "parameters": [
          {
            "name": "newPassword",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Yangi parol"
          }
        ],
        "responses": {
          "200": {
            "description": "Parol yangilandi"
          }
        }
      }
    },
    "/api/stores/month-foyda": {
      "put": {
        "summary": "Oylik foydani yangilash",
        "description": "Store ID token'dan otomatik alınır",
        "tags": ["Profit"],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "eski",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "yangi",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Oy foydasi yangilandi"
          }
        }
      },
      "get": {
        "summary": "Oylik foydani olish",
        "description": "Store ID token'dan otomatik alınır",
        "tags": ["Profit"],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Oylik foyda ro'yxati",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MonthFoyda"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/stores/products/{productId}/foyda": {
      "put": {
        "summary": "Mahsulot foydasini yangilash",
        "description": "Store ID token'dan otomatik alınır",
        "tags": ["Profit"],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "newValue",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mahsulot foydasi yangilandi"
          }
        }
      }
    },
    "/api/stores/products/stock": {
      "put": {
        "summary": "Mahsulot zahirasini yangilash",
        "description": "Store ID token'dan otomatik alınır",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "barcode",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "newStock",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mahsulot miqdori yangilandi"
          }
        }
      }
    },
    "/api/stores/foyda": {
      "get": {
        "summary": "Do'kon foydasini olish",
        "description": "Store ID token'dan otomatik alınır",
        "tags": ["Profit"],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Foyda miqdori",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    },
    "/api/sales": {
      "post": {
        "summary": "Yangi sotuv qo'shish",
        "description": "Store ID token'dan otomatik alınır",
        "tags": ["Sales"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Sales"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Muvaffaqiyatli qo'shildi"
          }
        }
      }
    },
    "/api/sales/{id}": {
      "get": {
        "summary": "Sotuvni olish",
        "tags": ["Sales"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sotuv ma'lumotlari",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sales"
                }
              }
            }
          }
        }
      }
    },
    "/api/stores/sales": {
      "get": {
        "summary": "Barcha sotuvlarni olish",
        "description": "Store ID token'dan otomatik alınır",
        "tags": ["Sales"],
        "responses": {
          "200": {
            "description": "Sotuvlar ro'yxati",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Sales"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/stores/sales/today": {
      "get": {
        "summary": "Bugungi sotuvlarni olish",
        "description": "Store ID token'dan otomatik alınır",
        "tags": ["Sales"],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bugungi sotuvlar ro'yxati",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Sales"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token ile kimlik doğrulama. Token'da store ID bilgisi bulunur."
      }
    },
    "schemas": {
      "AuthRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": {
            "type": "string",
            "example": "123",
            "description": "Store ID"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "Store password"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "description": "JWT token (store ID içerir)"
          }
        }
      },
      "ProductsObject": {
        "type": "object",
        "required": ["barcode", "name", "stock", "costPrice", "salePrice", "foyda"],
        "properties": {
          "productId": {
            "type": "integer",
            "example": 1
          },
          "storeId": {
            "type": "string",
            "example": "123",
            "description": "Token'dan otomatik alınır"
          },
          "barcode": {
            "type": "string",
            "example": "1234567890"
          },
          "name": {
            "type": "string",
            "example": "Pomidor"
          },
          "stock": {
            "type": "string",
            "example": "100"
          },
          "costPrice": {
            "type": "string",
            "example": "3000"
          },
          "salePrice": {
            "type": "string",
            "example": "3500"
          },
          "foyda": {
            "type": "string",
            "example": "500"
          },
          "monthFoyda": {
            "type": "integer",
            "example": 15000
          }
        }
      },
      "StoreDto": {
        "type": "object",
        "properties": {
          "storeId": {
            "type": "integer",
            "example": 123,
            "description": "Token'dan otomatik alınır"
          },
          "name": {
            "type": "string",
            "example": "SmartMarket"
          },
          "adress": {
            "type": "string",
            "example": "Toshkent, Chilonzor"
          },
          "email": {
            "type": "string",
            "example": "info@smartmarket.uz"
          },
          "number": {
            "type": "string",
            "example": "+998901234567"
          }
        }
      },
      "StoreUpdateDto": {
        "type": "object",
        "required": ["name", "email", "adress", "number", "password"],
        "properties": {
          "storeId": {
            "type": "integer",
            "example": 123,
            "description": "Token'dan otomatik alınır"
          },
          "name": {
            "type": "string",
            "example": "SmartMarket"
          },
          "email": {
            "type": "string",
            "example": "info@smartmarket.uz"
          },
          "adress": {
            "type": "string",
            "example": "Toshkent, Chilonzor"
          },
          "number": {
            "type": "string",
            "example": "+998901234567"
          },
          "password": {
            "type": "string",
            "example": "newpassword123"
          }
        }
      },
      "Sales": {
        "type": "object",
        "required": ["saleDate", "totalPrice", "productName", "barcode", "quantity", "isCash"],
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "storeId": {
            "type": "integer",
            "example": 123,
            "description": "Token'dan otomatik alınır"
          },
          "saleDate": {
            "type": "string",
            "format": "date-time",
            "example": "2025-01-15T10:30:00"
          },
          "totalPrice": {
            "type": "number",
            "format": "double",
            "example": 10500.0
          },
          "productName": {
            "type": "string",
            "example": "Pomidor"
          },
          "barcode": {
            "type": "string",
            "example": "1234567890"
          },
          "quantity": {
            "type": "integer",
            "example": 3
          },
          "isCash": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "MonthFoyda": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "storeId": {
            "type": "integer",
            "example": 123,
            "description": "Token'dan otomatik alınır"
          },
          "date": {
            "type": "string",
            "format": "date",
            "example": "2025-01-01"
          },
          "monthFoyda": {
            "type": "integer",
            "example": 150000
          }
        }
      }
    }
  }
}